Changed to the Windows 7.1 SDK
git-svn-id: https://g3d.svn.sourceforge.net/svnroot/g3d@1021 ef0428c2-0c16-0410-9f3f-d710ca1f8061

+++ G3D9/GLG3D.lib/source/GLCaps.cpp
@@ -153,7 +153,7 @@ std::string GLCaps::getDriverVersion() {
         RegistryUtil::readString(videoDeviceKey, format("%s%d", videoDeviceValue.c_str(), videoDeviceNum - 1), installedDriversKey);
 
         // find and remove the "\Registry\Machine\" part of the key
-        int subKeyIndex = installedDriversKey.find('\\', 1);
+        size_t subKeyIndex = installedDriversKey.find('\\', 1);
         subKeyIndex = installedDriversKey.find('\\', subKeyIndex + 1);
 
         installedDriversKey.erase(0, subKeyIndex);
